requirements.txt
monai
networkx
nibabel
numpy
pandas
pathos
pyyaml
scikit-image
scipy
SimpleITK
torch
tqdm
hydra-core>=1.3,<1.4